+++ archive/CA_DB_Config_2024-08-22.py
@@ -37,6 +37,7 @@ params_tables_fields_files = [Path(f).with_suffix('.csv') for f in
                               'recon',
                              ]]
 
+"""
 params_tables_fields = []
 for file in params_tables_fields_files:
     with open(file,'r') as f:
@@ -46,6 +47,7 @@ for file in params_tables_fields_files:
                 params_tables_fields.append(line)
             else:
                 pass
+"""
 
 metadata_fields = [
   'Dataset ID',
