+++ archive/frontend_prototype_archive/src/app/api/sample_scans/route.ts
