+++ archive/frontend_prototype_archive/src/app/globals.css
