+++ archive/frontend_prototype_archive/src/app/reconstruct/collapse_card.tsx
