+++ archive/frontend_prototype_archive/src/app/scans/page.tsx
